requirements.txt
flask
pandas
numpy
faiss-cpu
sentence-transformers
reportlab